+++ src/main/java/duke/DukeIllegalCommandException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DukeIllegalCommandException extends IllegalArgumentException {
     @Override
     public String toString() {
