+++ src/main/java/duke/Event.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.LocalDate;
 
 public class Event extends Task {
